requirements.txt
langchain
langchain-core
langchain-openai
langchain-pinecone
langchain-text-splitters
openai
pinecone
pydantic
python-dotenv
requests
tiktoken